+++ kitty/config.py
@@ -281,6 +281,7 @@ type_map = {
     'inactive_text_alpha': unit_float,
     'url_style': url_style,
     'prefer_color_emoji': to_bool,
+    'copy_on_select': to_bool,
 }
 
 for name in (