+++ CHANGELOG.rst
@@ -28,6 +28,9 @@ version 0.7.0 [future]
 - macOS: Fix closing top-level window does not transfer focus to other
   top-level windows.
 
+- macOS: Fix alt+arrow keys not working when disabling the macos_option_as_alt
+  config option.
+
 - kitty icat: Workaround for bug in ImageMagick that would cause some images
   to fail to display at certain sizes.